requirements.txt
# Core AI & API
google-genai>=1.0.0
elevenlabs>=2.0.0

# MCP (Model Context Protocol) for Playwright
mcp>=1.0.0
nest-asyncio>=1.6.0

# Speech & Audio
pvporcupine>=4.0.0
pvrecorder>=1.2.0
PyAudio>=0.2.14
SpeechRecognition>=3.10.0

# GUI & Screen Control
PyQt6>=6.0.0
pyautogui>=0.9.50
pillow>=10.0.0

# Environment & Utils
python-dotenv>=1.0.0
